fix: dcfgen doesn't generate SDO for setting PDO mapping to 0
If an EDS file doesn't have a default value for PDO mapping at sub index 0, dcfgen will not generate an SDO to reset
that PDO mapping to 0 before proceeding to setting the actual PDO mappings. Hence, instead of checking the value in
the EDS/DCF file, checking the mapping in the configuration fixes this problem.